requirements-light.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
WTForms==3.2.1
fpdf==1.7.2
Pillow==11.1.0
openpyxl==3.1.5
python-dateutil==2.9.0.post0
Jinja2==3.1.5
SQLAlchemy==2.0.37
Werkzeug==3.1.3
